Misc client edits
Design Print: There should be a space between the first and second
paragraph
-Case Studies portion text
Under Our Story: Please change this section
Creative Promotional Products is in the business of manufacturing and
selling products in promotional and retail markets.

To

For over 40 years, Creative Promotional Products has been in the
business of manufacturing and selling products in promotional and
retail markets.